+++ Framework/Interfaces/IDrawableComponent.cs
@@ -1,21 +1,9 @@
 ﻿namespace Macabre2D.Framework {
 
-    using System;
-
     /// <summary>
     /// Drawable component.
     /// </summary>
     public interface IDrawableComponent : IBaseComponent, IBoundable {
-
-        /// <summary>
-        /// Occurs when [draw order changed].
-        /// </summary>
-        event EventHandler DrawOrderChanged;
-
-        /// <summary>
-        /// Occurs when [visible changed].
-        /// </summary>
-        event EventHandler IsVisibleChanged;
 
         /// <summary>
         /// Gets the draw order.