+++ Framework/Interfaces/IUpdateableComponent.cs
@@ -6,6 +6,12 @@
     public interface IUpdateableComponent : IBaseComponent, IEnableableComponent {
 
         /// <summary>
+        /// Gets the update order.
+        /// </summary>
+        /// <value>The update order.</value>
+        int UpdateOrder { get; }
+
+        /// <summary>
         /// Updates this instance.
         /// </summary>
         /// <param name="frameTime">The frame time.</param>