+++ Framework/Interfaces/IUpdateableComponentAsync.cs
@@ -1,23 +1,11 @@
 ﻿namespace Macabre2D.Framework {
 
-    using System;
     using System.Threading.Tasks;
 
     /// <summary>
     /// An interface for a game component that can update on a background thread.
     /// </summary>
-    public interface IUpdateableComponentAsync : IBaseComponent {
-
-        /// <summary>
-        /// Occurs when [enabled changed].
-        /// </summary>
-        event EventHandler IsEnabledChanged;
-
-        /// <summary>
-        /// Gets a value indicating whether this <see cref="IUpdateableComponentAsync"/> is enabled.
-        /// </summary>
-        /// <value><c>true</c> if enabled; otherwise, <c>false</c>.</value>
-        bool IsEnabled { get; }
+    public interface IUpdateableComponentAsync : IBaseComponent, IEnableableComponent {
 
         /// <summary>
         /// Updates this instance asynchronously.