+++ Framework/Physics/SimplePhysicsModule.cs
@@ -16,11 +16,9 @@
         /// <value>The bodies.</value>
         protected FilterSortCollection<IPhysicsBody> Bodies { get; } = new FilterSortCollection<IPhysicsBody>(
             r => r.IsEnabled,
-            (r, handler) => r.IsEnabledChanged += handler,
-            (r, handler) => r.IsEnabledChanged -= handler,
+            nameof(IPhysicsBody.IsEnabled),
             (r1, r2) => Comparer<int>.Default.Compare(r1.UpdateOrder, r2.UpdateOrder),
-            (r, handler) => r.UpdateOrderChanged += handler,
-            (r, handler) => r.UpdateOrderChanged -= handler);
+            nameof(IPhysicsBody.UpdateOrder));
 
         /// <summary>
         /// Gets the collider tree.