+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Time handled through new TimeFrame object, which will allow adjustments to game speed.
 * Using MonoGame.WpfCore template instead of MonoGame.Framework.WpfInterop to display and edit scenes in the editor.
 * Updated icon to new simplified skull.
+* Components and modules use PropertyChanged instead of specific events.
 
 ### Fixes